requirements.txt
fastapi
uvicorn
python-dotenv
langchain
langgraph
langchain-anthropic
langchain-fireworks
langchain-google-genai
motor
pydantic
requests
black
isort
pytest
